SWINDyT/requirements.txt
torch>=2.0.0
torchvision>=0.15.0
torchmetrics>=0.11.0
numpy>=1.23.0
Pillow>=9.0.0
scikit-image>=0.20.0
scikit-learn>=1.2.0
opencv-python>=4.7.0
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
tqdm>=4.64.0
plotly>=5.0.0
lpips>=0.1.4
openpyxl>=3.1.0
torchsummary>=1.5.1
